+++ CHANGELOG.md
@@ -31,6 +31,9 @@
 
 * Viash namespace: Fix incorrect output path when the parent directory of a Viash component is not equal to the value of
   `.functionality.name`.
+  
+* `DockerPlatform`: Added sleep workaround to avoid concurrency issue where a file is executed to 
+  build docker containers but apparently still in the process of being written.
 
 # Viash 0.5.4